Move server/mysql_quota_provider.go into quota/mysqlqm package
Makes it optional to import.
