+++ cmd/trillian_log_server/main.go
@@ -56,6 +56,9 @@ import (
 
 	// Load hashers
 	_ "github.com/google/trillian/merkle/rfc6962"
+
+	// Load MySQL quota provider
+	_ "github.com/google/trillian/quota/mysqlqm"
 )
 
 var (
