+++ cmd/trillian_log_signer/main.go
@@ -57,6 +57,9 @@ import (
 
 	// Load hashers
 	_ "github.com/google/trillian/merkle/rfc6962"
+
+	// Load MySQL quota provider
+	_ "github.com/google/trillian/quota/mysqlqm"
 )
 
 var (
